requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
google-genai>=1.0.0
google-adk>=1.0.0
